+++ CHANGELOG.md
@@ -11,6 +11,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - [#674](http://github.com/nelson-lang/nelson/issues/674) Migrate sources to nelson-lang github organization.
 - `cellfun` did not check type of second input argument.
 
+### Added
+
+- control system module (part 2):
+  - `nyquist` Nyquist plot of frequency response.
+
 ### Changed
 
 - [#674](http://github.com/nelson-lang/nelson/issues/674) Migrate sources to nelson-lang github organization.